31 July, Wednesday - 1807
- added the codes required for multiple select to work in the add() in
PdsController.
- removed the comment on line 55 $this.->redirect(array('action' =>
'index')); in PdsController
- added validation for audience in the Pd Model. There is an issue
where on edit, if the audience is left blank, the page redirect goes to
a blank page. (Upon edit, the values stored in the db.audience doesn't
seem to be written or be displayed (line 7 in Pd.php).
- made the changes in multiple selects after the changes in the
PdsController has been done (line 31-33, line 42-54 in add.ctp).
- added in-line markup to indicate that the Intended Audience field is
mandatory. This is in relation to the bug that needs fixing (line 33 in
edit.ctp).
- changed Targeted Audience to Intended Audience for consistency (line
9 in view.ctp).